+++ lib/runTest.ts
@@ -4,9 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as cp from 'child_process';
-import * as path from 'path';
-import { DownloadOptions, defaultCachePath, downloadAndUnzipVSCode } from './download';
-import { killTree } from './util';
+import { DownloadOptions, downloadAndUnzipVSCode } from './download';
+import { getProfileArguments, killTree } from './util';
 
 export interface TestOptions extends Partial<DownloadOptions> {
 	/**
@@ -108,25 +107,6 @@ export async function runTests(options: TestOptions): Promise<number> {
 
 	return innerRunTests(options.vscodeExecutablePath, args, options.extensionTestsEnv);
 }
-
-/** Adds the extensions and user data dir to the arguments for the VS Code CLI */
-export function getProfileArguments(args: readonly string[]) {
-	const out: string[] = [];
-	if (!hasArg('extensions-dir', args)) {
-		out.push(`--extensions-dir=${path.join(defaultCachePath, 'extensions')}`);
-	}
-
-	if (!hasArg('user-data-dir', args)) {
-		out.push(`--user-data-dir=${path.join(defaultCachePath, 'user-data')}`);
-	}
-
-	return out;
-}
-
-function hasArg(argName: string, argList: readonly string[]) {
-	return argList.some((a) => a === `--${argName}` || a.startsWith(`--${argName}=`));
-}
-
 const SIGINT = 'SIGINT';
 
 async function innerRunTests(