+++ vitest.config.ts
@@ -6,8 +6,8 @@
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  test: {
-		include: ['lib/**/*.test.ts'],
+	test: {
+		include: ['lib/**/*.test.{ts,mts}'],
 		testTimeout: 120_000,
-  },
+	},
 });