+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow installing unreleased builds using an `-unreleased` suffix, such as `insiders-unreleased`.
 - Allow passing different data directories in `runVSCodeCommand`, using it for extension development.
+- Improve the appearance progress reporting.
 
 ### 2.3.10 | 2024-05-13